Refactor scripts to be split into  one file per operation
